requirements.txt
anytree
asyncio
beautifulsoup4
bencoding
cfscrape
dnspython
feedparser
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
git+https://github.com/zevtyardt/lk21.git
lxml
pillow
psutil
pybase64
pymongo
pyrogram==2.0.66
python-dotenv
python-magic
python-telegram-bot==13.15
requests
telegraph
tenacity
tgCrypto
xattr
beautifulsoup4
bencode.py
bencoding
cfscrape